主キーなし

+++ samplecom/src/main/java/com/example/action/model/base/T00UniqueSDeleteAction.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import com.example.entity.T00Nokey;
+import com.example.entity.T00Unique;
 
 import jp.co.golorp.emarf.action.BaseAction;
 import jp.co.golorp.emarf.exception.OptLockError;
@@ -13,13 +13,13 @@ import jp.co.golorp.emarf.util.Messages;
 import jp.co.golorp.emarf.validation.FormValidator;
 
 /**
- * キーなし一覧削除
+ * ユニークキー一覧削除
  *
  * @author emarfkrow
  */
-public class T00NokeySDeleteAction extends BaseAction {
+public class T00UniqueSDeleteAction extends BaseAction {
 
-    /** キーなし一覧削除処理 */
+    /** ユニークキー一覧削除処理 */
     @Override
     public Map<String, Object> running(final LocalDateTime now, final String execId, final Map<String, Object> form) {
 
@@ -28,7 +28,7 @@ public class T00NokeySDeleteAction extends BaseAction {
         int count = 0;
 
         @SuppressWarnings("unchecked")
-        List<Map<String, Object>> data = (List<Map<String, Object>>) form.get("T00NokeyGrid");
+        List<Map<String, Object>> data = (List<Map<String, Object>>) form.get("T00UniqueGrid");
         if (data != null) {
             for (Map<String, Object> row : data) {
 
@@ -37,10 +37,16 @@ public class T00NokeySDeleteAction extends BaseAction {
                 }
 
                 // 主キーが不足していたらエラー
+                if (jp.co.golorp.emarf.lang.StringUtil.isNullOrWhiteSpace(row.get("D_MEI"))) {
+                    throw new OptLockError("error.cant.delete", "ユニークキー");
+                }
+                if (jp.co.golorp.emarf.lang.StringUtil.isNullOrWhiteSpace(row.get("E_MEI"))) {
+                    throw new OptLockError("error.cant.delete", "ユニークキー");
+                }
 
-                T00Nokey e = FormValidator.toBean(T00Nokey.class.getName(), row);
+                T00Unique e = FormValidator.toBean(T00Unique.class.getName(), row);
                 if (e.delete() != 1) {
-                    throw new OptLockError("error.cant.delete", "キーなし");
+                    throw new OptLockError("error.cant.delete", "ユニークキー");
                 }
                 ++count;
             }
